refactor: cria diretório próprio para as tabelas do AS
