--- electron/database/plunder.ts
@@ -1,486 +1,0 @@
-import { DataTypes, Model } from 'sequelize';
-import { assertInteger } from '$global/guards';
-import { sequelize } from '$electron/database';
-import { isUserAlias, assertUserAlias, assertWallLevel } from '$global/guards';
-import { DatabaseError } from '$electron/error';
-import { unitsToDestroyWall } from '$global/constants';
-import type { InferAttributes, InferCreationAttributes, CreationOptional } from 'sequelize';
-import type { UserAlias } from '$types/electron';
-import type { useCacheStore } from '$electron/interface';
-import type { WallLevel, UnitsToDestroyWall } from '$types/game';
-
-import type {
-    PlunderConfigType,
-    PlunderHistoryType,
-    PlunderAttackDetails,
-    BlindAttackPattern,
-    UseCPattern,
-    CustomPlunderTemplateType,
-    DemolitionTemplateType
-} from '$types/plunder';
-
-export class PlunderConfig extends Model<InferAttributes<PlunderConfig>, InferCreationAttributes<PlunderConfig>> implements PlunderConfigType {
-    declare readonly id: UserAlias;
-
-    // Painel
-    declare readonly active: boolean;
-    declare readonly ignoreWall: boolean;
-    declare readonly destroyWall: boolean;
-    declare readonly groupAttack: boolean;
-    declare readonly useC: boolean;
-    declare readonly ignoreDelay: boolean;
-    declare readonly blindAttack: boolean;
-
-    // Ataque
-    declare readonly maxDistance: number;
-    declare readonly ignoreOlderThan: number;
-    declare readonly attackDelay: number;
-    declare readonly resourceRatio: number;
-    declare readonly blindAttackPattern: BlindAttackPattern;
-
-    // Modelo C
-    declare readonly useCPattern: UseCPattern;
-    declare readonly maxDistanceC: number;
-    declare readonly ignoreOlderThanC: number;
-    declare readonly useCWhenResourceRatioIsBiggerThan: number;
-
-    // Grupo
-    declare readonly plunderGroupId: number | null;
-    declare readonly fieldsPerWave: number;
-    declare readonly villageDelay: number;
-
-    // Muralha
-    declare readonly wallLevelToIgnore: WallLevel;
-    declare readonly wallLevelToDestroy: WallLevel;
-    declare readonly destroyWallMaxDistance: number;
-    
-    // Outros
-    declare readonly minutesUntilReload: number;
-    declare readonly plunderedResourcesRatio: number;
-    declare readonly pageDelay: number;
-};
-
-PlunderConfig.init({
-    id: {
-        type: DataTypes.STRING,
-        primaryKey: true,
-        allowNull: false,
-        unique: true,
-        validate: {
-            isUserAlias(value: unknown) {
-                assertUserAlias(value, DatabaseError);
-            }
-        }
-    },
-
-    // Painel
-    active: {
-        type: DataTypes.BOOLEAN,
-        allowNull: false,
-        defaultValue: false
-    },
-    ignoreWall: {
-        type: DataTypes.BOOLEAN,
-        allowNull: false,
-        defaultValue: false
-    },
-    destroyWall: {
-        type: DataTypes.BOOLEAN,
-        allowNull: false,
-        defaultValue: false
-    },
-    groupAttack: {
-        type: DataTypes.BOOLEAN,
-        allowNull: false,
-        defaultValue: false
-    },
-    useC: {
-        type: DataTypes.BOOLEAN,
-        allowNull: false,
-        defaultValue: false
-    },
-    ignoreDelay: {
-        type: DataTypes.BOOLEAN,
-        allowNull: false,
-        defaultValue: false
-    },
-    blindAttack: {
-        type: DataTypes.BOOLEAN,
-        allowNull: false,
-        defaultValue: false
-    },
-
-    // Ataque
-    maxDistance: {
-        type: DataTypes.FLOAT,
-        allowNull: false,
-        defaultValue: 20,
-        validate: {
-            min: 1,
-            isFloat: true
-        }
-    },
-    ignoreOlderThan: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        defaultValue: 10,
-        validate: {
-            isInt: true
-        }
-    },
-    attackDelay: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        defaultValue: 200,
-        validate: {
-            min: 100,
-            max: 60000,
-            isInt: true
-        }
-    },
-    resourceRatio: {
-        type: DataTypes.FLOAT,
-        allowNull: false,
-        defaultValue: 0.8,
-        validate: {
-            min: 0.2,
-            max: 1,
-            isFloat: true
-        }
-    },
-    blindAttackPattern: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        defaultValue: 'smaller' satisfies BlindAttackPattern,
-        validate: {
-            isIn: [['smaller', 'bigger'] satisfies BlindAttackPattern[]]
-        }
-    },
-
-    // Modelo C
-    useCPattern: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        defaultValue: 'normal' satisfies UseCPattern,
-        validate: {
-            isIn: [['excess', 'normal', 'only'] satisfies UseCPattern[]]
-        }
-    },
-    maxDistanceC: {
-        type: DataTypes.FLOAT,
-        allowNull: false,
-        defaultValue: 10,
-        validate: {
-            min: 1,
-            isFloat: true
-        }
-    },
-    ignoreOlderThanC: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        defaultValue: 5,
-        validate: {
-            isInt: true
-        }
-    },
-    useCWhenResourceRatioIsBiggerThan: {
-        type: DataTypes.FLOAT,
-        allowNull: false,
-        defaultValue: 3,
-        validate: {
-            min: 1,
-            isFloat: true
-        }
-    },
-
-    // Grupo
-    plunderGroupId: {
-        type: DataTypes.INTEGER,
-        allowNull: true
-    },
-    fieldsPerWave: {
-        type: DataTypes.FLOAT,
-        allowNull: false,
-        defaultValue: 10,
-        validate: {
-            min: 5,
-            isFloat: true
-        }
-    },
-    villageDelay: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        defaultValue: 2000,
-        validate: {
-            min: 100,
-            max: 60000,
-            isInt: true
-        }
-    },
-
-    // Muralha
-    wallLevelToIgnore: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        defaultValue: 1,
-        validate: {
-            isWallLevel(value: unknown) {
-                assertWallLevel(value, DatabaseError);
-            }
-        }
-    },
-    
-    wallLevelToDestroy: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        defaultValue: 1,
-        validate: {
-            isWallLevel(value: unknown) {
-                assertWallLevel(value, DatabaseError);
-            }
-        }
-    },
-    destroyWallMaxDistance: {
-        type: DataTypes.FLOAT,
-        allowNull: false,
-        defaultValue: 20,
-        validate: {
-            min: 1,
-            isFloat: true
-        }
-    },
-
-    // Outros
-    minutesUntilReload: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        defaultValue: 10,
-        validate: {
-            min: 1,
-            max: 60,
-            isInt: true
-        }
-    },
-    plunderedResourcesRatio: {
-        type: DataTypes.FLOAT,
-        allowNull: false,
-        defaultValue: 1,
-        validate: {
-            min: 0.2,
-            max: 1,
-            isFloat: true
-        }
-    },
-    pageDelay: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        defaultValue: 2000,
-        validate: {
-            min: 100,
-            max: 60000,
-            isInt: true
-        }
-    }  
-}, { sequelize, tableName: 'plunder_config', timestamps: true });
-
-export class PlunderHistory extends Model<InferAttributes<PlunderHistory>, InferCreationAttributes<PlunderHistory>> implements PlunderHistoryType {
-    declare readonly id: UserAlias;
-    declare readonly last: PlunderAttackDetails;
-    declare readonly total: PlunderAttackDetails;
-
-    public static async getHistoryAsJSON(cacheStore: ReturnType<typeof useCacheStore>): Promise<PlunderHistoryType | null> {
-        try {
-            const userAlias = cacheStore.userAlias;
-            if (!isUserAlias(userAlias)) return null;
-
-            const plunderHistory = await PlunderHistory.findByPk(userAlias);
-            return plunderHistory ? plunderHistory.toJSON() : null;
-    
-        } catch (err) {
-            DatabaseError.catch(err);
-            return null;
-        };
-    }
-};
-
-PlunderHistory.init({
-    id: {
-        type: DataTypes.STRING,
-        primaryKey: true,
-        allowNull: false,
-        unique: true,
-        validate: {
-            isUserAlias(value: unknown) {
-                assertUserAlias(value, DatabaseError);
-            }
-        }
-    },
-    last: {
-        type: DataTypes.JSON,
-        allowNull: false
-    },
-    total: {
-        type: DataTypes.JSON,
-        allowNull: false
-    }
-}, { sequelize, tableName: 'plunder_history', timestamps: true });
-
-export class CustomPlunderTemplate extends Model<
-    InferAttributes<CustomPlunderTemplate>,
-    InferCreationAttributes<CustomPlunderTemplate>>
-implements CustomPlunderTemplateType {
-    declare readonly id: CreationOptional<number>;
-    declare readonly alias: UserAlias;
-    declare readonly type: string;
-    declare readonly description: string | null;
-    declare readonly units: CustomPlunderTemplateType['units'];
-
-    public static async getCustomPlunderTemplates(alias: UserAlias): Promise<CustomPlunderTemplateType[] | null> {
-        try {
-            assertUserAlias(alias, DatabaseError);
-            const templates = await CustomPlunderTemplate.findAll({ where: { alias } });
-            return templates.map((template) => template.toJSON());
-
-        } catch (err) {
-            DatabaseError.catch(err);
-            return null;
-        };
-    };
-
-    public static async saveCustomPlunderTemplate(template: CustomPlunderTemplateType): Promise<boolean> {
-        try {
-            for (const [unit, amount] of Object.entries(template.units)) {
-                assertInteger(amount, `Invalid ${unit} amount: ${amount}`);
-                if (amount < 0) throw new DatabaseError(`${unit} amount cannot be negative`);
-            };
-
-            await sequelize.transaction(async (transaction) => {
-                await CustomPlunderTemplate.upsert({ ...template }, { transaction });
-            });
-
-            return true;
-
-        } catch (err) {
-            DatabaseError.catch(err);
-            return false;
-        };
-    };
-
-    public static async destroyCustomPlunderTemplate(template: CustomPlunderTemplateType): Promise<boolean> {
-        try {
-            assertUserAlias(template.alias, DatabaseError);
-            await sequelize.transaction(async (transaction) => {
-                await CustomPlunderTemplate.destroy({ where: { alias: template.alias, type: template.type }, transaction });
-            });
-
-            return true;
-
-        } catch (err) {
-            DatabaseError.catch(err);
-            return false;
-        };
-    };
-};
-
-CustomPlunderTemplate.init({
-    id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        autoIncrement: true
-    },
-    alias: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        validate: {
-            isUserAlias(value: unknown) {
-                assertUserAlias(value, DatabaseError);
-            }
-        }
-    },
-    type: {
-        type: DataTypes.STRING,
-        allowNull: false
-    },
-    description: {
-        type: DataTypes.STRING,
-        allowNull: true
-    },
-    units: {
-        type: DataTypes.JSON,
-        allowNull: false
-    }
-}, { sequelize, tableName: 'custom_plunder_template', timestamps: true });
-
-export class DemolitionTemplate extends Model<
-    InferAttributes<DemolitionTemplate>,
-    InferCreationAttributes<DemolitionTemplate>>
-implements DemolitionTemplateType {
-    declare readonly id: CreationOptional<number>;
-    declare readonly alias: UserAlias;
-    declare readonly units: UnitsToDestroyWall;
-
-    public static async getDemolitionTroopsConfig(alias: UserAlias): Promise<DemolitionTemplateType | null> {
-        try {
-            assertUserAlias(alias, DatabaseError);
-            const demolitionTemplate = await DemolitionTemplate.findOne({ where: { alias } });
-            if (!demolitionTemplate) {
-                return { alias, units: unitsToDestroyWall };
-            };
-            
-            return demolitionTemplate.toJSON();
-
-        } catch (err) {
-            DatabaseError.catch(err);
-            return null;
-        };
-    };
-
-    public static async saveDemolitionTroopsConfig(template: DemolitionTemplateType): Promise<boolean> {
-        try {
-            await sequelize.transaction(async (transaction) => {
-                await DemolitionTemplate.upsert({ ...template }, { transaction });
-            });
-            return true;
-
-        } catch (err) {
-            DatabaseError.catch(err);
-            return false;
-        };
-    };
-
-    public static async destroyDemolitionTroopsConfig(alias: UserAlias): Promise<boolean> {
-        try {
-            assertUserAlias(alias, DatabaseError);
-            await sequelize.transaction(async (transaction) => {
-                await DemolitionTemplate.destroy({ where: { alias }, transaction });
-            });
-            return true;
-
-        } catch (err) {
-            DatabaseError.catch(err);
-            return false;
-        };
-    };
-};
-
-DemolitionTemplate.init({
-    id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        autoIncrement: true
-    },
-    alias: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        unique: true,
-        validate: {
-            isUserAlias(value: unknown) {
-                assertUserAlias(value, DatabaseError);
-            }
-        }
-    },
-    units: {
-        type: DataTypes.JSON,
-        allowNull: false
-    }
-}, { sequelize, tableName: 'demolition_template', timestamps: true });
