+++ electron/database/plunder/index.ts
@@ -1,0 +1,4 @@
+export * from '$database/plunder/config';
+export * from '$database/plunder/demolition';
+export * from '$database/plunder/history';
+export * from '$database/plunder/template';
